# openapi.yaml
openapi: 3.0.2
info:
  description: Moov ACH ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating ACH files. ACH is the primary method of electronic money movement throughout the United States.
  version: v1
  title: ACH API
  contact:
    url: https://github.com/Alviere/ach
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /files:
    get:
      tags: ['ACH Files']
      summary: List Files
      description: List all ACH files created with the ACH service. These files are not persisted through multiple runs of the service.
      operationId: getFiles
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
      responses:
        '200':
          description: A list of File objects
          headers:
            X-Total-Count:
              description: The total number of Originators
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Files'
  /files/create:
    post:
      tags: ['ACH Files']
      summary: Create File
      description: Create a new File object from either the plaintext or JSON representation.
      operationId: createFile
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Idempotent key in the header which expires after 24 hours. These strings should contain enough entropy for to not collide with each other in your requests.
          example: "a4f88150"
          required: false
          schema:
            type: string
        - name: requireABAOrigin
          in: query
          description: Optional parameter to enable routing number validation over ImmediateOrigin
          schema:
            $ref: '#/components/schemas/ValidateOpts/properties/requireABAOrigin'
        - name: bypassOrigin
          in: query
          description: Optional parameter to disable ImmediateOrigin validation
          schema:
            $ref: '#/components/schemas/ValidateOpts/properties/bypassOriginValidation'
        - name: bypassDestination
          in: query
          description: Optional parameter to disable ImmediateDestination validation
          schema:
            $ref: '#/components/schemas/ValidateOpts/properties/bypassDestinationValidation'
      requestBody:
        description: Content of the ACH file (in json or raw text)
        required: true
        content:
          text/plain:
            schema:
              description: A plaintext ACH file
              type: string
              example: |
                101 23138010401210428821906240000A094101Federal Reserve Bank   My Bank Name
                5225Name on Account                     121042882 PPDREG.SALARY      190625   1121042880000001
                62723138010412345678         0100000000               Receiver Account Name   0121042880000001
                82250000010023138010000100000000000000000000121042882                          121042880000001
                9000001000001000000010023138010000100000000000000000000
                9999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999
                9999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999
                9999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999
                9999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999
                9999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFile'
      responses:
        '200':
          description: A JSON object containing a new File
          headers:
            Location:
              description: The location of the new resource
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileID'
        '400':
          description: "Invalid File Header Object"
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
  /files/{fileID}:
    get:
      tags: ['ACH Files']
      summary: Retrieve File
      description: Get the details of an existing File using the unique File identifier that was returned upon creation.
      operationId: getFileByID
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      responses:
        '200':
          description: A File object for the supplied ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '404':
          description: A resource with the specified ID was not found
    delete:
      tags: ['ACH Files']
      summary: Delete File
      description: Permanently deletes a File and associated Batches. It cannot be undone.
      operationId: deleteACHFile
      parameters:
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
      responses:
        '200':
          description: Permanently deleted File.
        '404':
          description: A File with the specified ID was not found.
  /files/{fileID}/contents:
    get:
      tags: ['ACH Files']
      summary: Get File Contents
      description: |
        Assembles the existing file (batches and controls) records, computes sequence numbers and totals. Returns plaintext file.
      operationId: getFileContents
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      responses:
        '200':
          description: File built successfully without errors.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RawFile'
  /files/{fileID}/validate:
    get:
      tags: ['ACH Files']
      summary: Validate File
      description: Validates the existing file. You need only supply the unique File identifier that was returned upon creation.
      operationId: checkFile
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      responses:
        '200':
          description: File validated successfully without errors.
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
        '400':
          description: Validation failed. Check response for errors
    post:
      tags: ['ACH Files']
      summary: Validate File (Custom)
      description: Validates the existing file. You need only supply the unique File identifier that was returned upon creation.
      operationId: validateFile
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateOpts'
      responses:
        '200':
          description: File validated successfully without errors.
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
        '400':
          description: Validation failed. Check response for errors
  /files/{fileID}/segment:
    post:
      tags: ['ACH Files']
      summary: Segment File
      description: Split one file into two. One with only debits and one with only credits.
      operationId: segmentFile
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Idempotent key in the header which expires after 24 hours. These strings should contain enough entropy for to not collide with each other in your requests.
          example: "a4f88150"
          required: false
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      # requestBody:
      #   description: Optional configuration for segmenting files
      #   required: false
      #   content:
      #     application/json:
      #       schema:
      #         $ref: '#/components/schemas/SegmentFileConfiguration'
      responses:
        '200':
          description: An ID of the new ACH file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SegmentedFiles'
        '400':
          description: See error in response body
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
  /files/{fileID}/flatten:
    post:
      tags: ['ACH Files']
      summary: Flatten Batches
      description: Consolidate batches and entries into the minimum number of batches needed.
      operationId: flattenFile
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Idempotent key in the header which expires after 24 hours. These strings should contain enough entropy for to not collide with each other in your requests.
          example: "a4f88150"
          required: false
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      responses:
        '200':
          description: An ID of the new ACH file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileID'
        '400':
          description: See error in response body
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
  /files/{fileID}/batches:
    get:
      tags: ['ACH Files']
      summary: Get Batches
      description: Get the batches on a File.
      operationId: getFileBatches
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      responses:
        '200':
          description: A list of Batch objects
          headers:
            X-Total-Count:
              description: The total number of Batches on the File.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batches'
    post:
      tags: ['ACH Files']
      summary: Append Batch to File
      description: Append a Batch record to the specified file
      operationId: addBatchToFile
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Idempotent key in the header which expires after 24 hours. These strings should contain enough entropy for to not collide with each other in your requests.
          example: "a4f88150"
          required: false
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Batch'
      responses:
        '200':
          description: Batch added to File
  /files/{fileID}/batches/{batchID}:
    get:
      tags: ['ACH Files']
      summary: Get Batch
      description: Get a specific Batch on a File
      operationId: getFileBatch
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
        - name: batchID
          in: path
          description: Batch ID
          required: true
          schema:
            type: string
            example: "45758063"
      responses:
        '200':
          description: Batch object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '404':
          description: Batch or File not found
    delete:
      tags: ['ACH Files']
      summary: Delete Batch
      description: Delete a Batch from a File
      operationId: deleteFileBatch
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
        - name: batchID
          in: path
          description: Batch ID
          required: true
          schema:
            type: string
            example: "45758063"
      responses:
        '200':
          description: Batch deleted
        '404':
          description: Batch or File not found
components:
  schemas:
    CreateFile:
      properties:
        ID:
          type: string
          description: File ID
          example: "3f2d23ee214"
        fileHeader:
          $ref: '#/components/schemas/FileHeader'
        batches:
          type: array
          items:
            $ref: '#/components/schemas/Batch'
        IATBatches:
          type: array
          items:
            $ref: '#/components/schemas/IATBatch'
        fileControl:
          $ref: '#/components/schemas/FileControl'
        advEntryDetails:
          type: array
          items:
            $ref: '#/components/schemas/ADVEntryDetail'
        advBatchControl:
          $ref: '#/components/schemas/ADVBatchControl'
      required:
        - fileHeader
    FileID:
      properties:
        ID:
          type: string
          description: File ID
          example: "1e522dc8"
        error:
          type: string
          description: An error message describing the problem intended for humans.
          example: Validation error(s) present.
    File:
      properties:
        ID:
          type: string
          description: File ID
          example: "3f2d23ee214"
        fileHeader:
          $ref: '#/components/schemas/FileHeader'
        batches:
          type: array
          items:
            $ref: '#/components/schemas/Batch'
        IATBatches:
          type: array
          items:
            $ref: '#/components/schemas/IATBatch'
        fileControl:
          $ref: '#/components/schemas/FileControl'
        NotificationOfChange:
          type: array
          items:
            $ref: '#/components/schemas/Batch'
          nullable: true
        ReturnEntries:
          type: array
          items:
            $ref: '#/components/schemas/Batch'
          nullable: true
      required:
        - ID
        - fileHeader
        - fileControl
    FileHeader:
      properties:
        immediateOrigin:
          type: string
          description: contains the Routing Number of the ACH Operator or sending point that is sending the file.
          minLength: 9
          maxLength: 10
          example: "99991234"
        immediateOriginName:
          type: string
          description: The name of the ACH operator or sending point that is sending the file.
          maxLength: 23
          example: "My Bank Name"
        immediateDestination:
          type: string
          maxLength: 10
          minLength: 9
          example: "69100013"
          description: contains the Routing Number of the ACH Operator or receiving point to which the file is being sent
        immediateDestinationName:
          type: string
          description: The name of the ACH or receiving point for which that file is destined.
          maxLength: 23
          example: "Federal Reserve Bank"
        fileCreationTime:
          description: 'The File Creation Date is the date when the file was prepared by an ODFI. (Format HHmm - H=Hour, m=Minute)'
          type: string
          format: "HHmm"
          example: "1504"
        fileCreationDate:
          description: 'The File Creation Time is the time when the file was prepared by an ODFI. (Format YYMMDD - Y=Year, M=Month, D=Day)'
          type: string
          format: "YYMMDD"
          example: "190102"
        fileIDModifier:
          type: string
          description: Incremented value for each file for RDFI's.
          example: "0"
      required:
        - immediateOrigin
        - immediateOriginName
        - immediateDestination
        - immediateDestinationName
        - fileCreationTime
        - fileCreationDate
    FileControl:
      properties:
        ID:
          description: Moov API File ID
          type: string
          example: "d1e26288"
        batchCount:
          description: Count of Batches in the File
          type: integer
          minimum: 1
          example: 1
        blockCount:
          description: |
            BlockCount total number of records in the file (include all headers and trailer) divided by 10 (This number must be evenly divisible by 10. If not, additional records consisting of all 9's are added to the file after the initial '9' record to fill out the block 10.)
          type: integer
          example: 1
        entryAddendaCount:
          description: Total detail and addenda records in the file
          type: integer
          minimum: 1
          example: 1
        entryHash:
          description: EntryHash calculated in the same manner as the batch total but includes total from entire file
          type: integer
          example: 0
        totalDebit:
          description: Accumulated Batch debit totals within the file.
          type: integer
          example: 100
        totalCredit:
          description: Accumulated Batch credit totals within the file.
          type: integer
          example: 20
      required:
        - ID 
        - batchCount
        - blockCount
        - entryAddendaCount
        - entryHash
        - totalDebit
        - totalCredit
    RawFile:
      type: string
      description: Plaintext ACH file
      example: "101 222380104 1210428821805100000A094101Citadel                Bank Name"
    Files:
      type: array
      items:
        $ref: '#/components/schemas/File'
    Batch:
      properties:
        batchHeader:
          $ref: '#/components/schemas/BatchHeader'
        entryDetails:
          type: array
          items:
            $ref: '#/components/schemas/EntryDetail'
          minLength: 1
        batchControl:
          $ref: '#/components/schemas/BatchControl'
        offset:
          $ref: '#/components/schemas/Offset'
      required:
        - batchHeader
        - entryDetails
        - batchControl
        - offset
...
    Batches:
      type: array
      items:
        $ref: '#/components/schemas/Batch'
...
    IATBatch:
      properties:
        ID:
          description: Client defined string used as a reference to this record.
          type: string
          example: a747e53f
        IATBatchHeader:
          $ref: '#/components/schemas/IATBatchHeader'
        IATEntryDetails:
          type: array
          items:
            $ref: '#/components/schemas/IATEntryDetail'
          minItems: 1
        batchControl:
          $ref: '#/components/schemas/BatchControl'
      required:
         - IATBatchHeader
         - IATEntryDetails
         - batchControl
...
    Offset:
      properties:
        routingNumber:
          type: string
          description: ABA routing number
          example: "987654320"
        accountNumber:
          type: string
          description: Account number used to offset records
        accountType:
          type: string
          description: Account type used in offset record
          enum:
            - checking
            - savings
        description:
          type: string
          description: Memo for Offset EntryDetail record
          example: OFFSET
      required:
        - routingNumber 
        - accountNumber
        - accountType
        - description
    SegmentedFiles:
      properties:
        creditFileID:
          type: string
          description: File ID
          example: "058960d8"
        debitFileID:
          type: string
          description: File ID
          example: "3cac5447"

    # There are no options here currently, but this object exists to read them in the future
    # SegmentFileConfiguration:
    #   properties: {}
    ValidateOpts:
      properties:
        requireABAOrigin:
          type: boolean
          default: false
          description: Require that the FileHeader ImmediateOrigin routing number which checksum matches.
        bypassOriginValidation:
          type: boolean
          default: false
          description: Skip ImmediateOrigin validation steps.
        bypassDestinationValidation:
          type: boolean
          default: false
          description: Skip ImmediateDestination validation steps.